# bluebell/docs/swagger.yaml
basePath: /api/v1
definitions:
  controller._ResponsePostList:
    properties:
      code:
        description: 业务响应状态码
        type: integer
      data:
        description: 数据
        items:
          $ref: '#/definitions/models.ApiPostDetail'
        type: array
      message:
        description: 提示信息
        type: string
    type: object
  models.ApiPostDetail:
    properties:
      author_id:
        type: integer
      authorName:
        type: string
      community_id:
        type: integer
      content:
        type: string
      create_time:
        type: string
      id:
        type: integer
      introduction:
        type: string
      name:
        type: string
      status:
        type: integer
      title:
        type: string
      vote_num:
        type: integer
    required:
    - community_id
    - content
    - title
    type: object
host: localhost:8080
info:
  contact:
    name: zhou
    url: https://www.fishworm96.github.io
  description: Go web开发进阶项目实战bluebell
  title: bluebell项目接口文档
  version: "1.0"
paths:
  /post:
    post:
      consumes:
      - application/json
      description: 可按社区按时间或分数排序查询帖子列表接口2
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 可以为空
        in: query
        name: community_id
        type: integer
      - description: 排序依据
        example: score
        in: query
        name: order
        type: string
      - description: 页码
        example: 1
        in: query
        name: page
        type: integer
      - description: 每页数据量
        example: 10
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      security:
      - ApiKeyAuth: []
      summary: 根据id获取帖子信息
      tags:
      - 帖子相关接口(api分组展示使用的)
  /posts2:
    get:
      consumes:
      - application/json
      description: 可按社区按时间或分数排序查询帖子列表接口
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 可以为空
        in: query
        name: community_id
        type: integer
      - description: 排序依据
        example: score
        in: query
        name: order
        type: string
      - description: 页码
        example: 1
        in: query
        name: page
        type: integer
      - description: 每页数据量
        example: 10
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      security:
      - ApiKeyAuth: []
      summary: 升级版帖子列表接口
      tags:
      - 帖子相关接口(api分组展示使用的)
swagger: "2.0"